Launching of runners added to StackCoordinator and related changes on several others

+++ spi/src/main/java/org/apache/usergrid/chop/spi/InstanceManager.java
@@ -13,7 +13,7 @@ import org.apache.usergrid.chop.stack.InstanceSpec;
  * Manages instances.
  */
 public interface InstanceManager {
-    int getDefaultTimeout();
+    long getDefaultTimeout();
 
     void terminateInstances( Collection<String> instancesIds );
